cgma-igm-comparison-ui-app/python-backend/requirements.txt
fastapi==0.115.*
uvicorn[standard]==0.34.*
rdflib==7.*
requests==2.*
python-dotenv==1.*
pytest==8.*
httpx==0.28.*
